pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Recallable Python project template"
requires-python = ">=3.10"

dependencies = [
  "pyyaml",
  "tenacity",
  "loguru"
]

[tool.setuptools.packages.find]
where = ["src"]
